Phase 2 Refactoring: Folder reorganization and cleanup
- Moved NewUI/Scripts/* to Scripts/UI/ (consolidated UI scripts)
- Moved NewUI assets (USS, UXML, Prefabs) to UI/ folder
- Deleted legacy uGUI scripts (Controllers/, Services/, etc.)
- Deleted unused prefabs (GridCellUI, PlayerGridPanel, etc.)
- Deleted test scenes (GuillotineTesting, NewPlayTesting, NetworkingTest)
- Created AudioSettings.cs for volume helper methods
- Added WordPlacementData.RowIndex and GameplayStateTracker.AddOpponentMisses
- Fixed DifficultySetting casts in opponent classes
- Cleaned up scene (removed missing script GameObjects)
- Removed development comments from UI scripts
- Marked Phase 2 refactoring plan as FINAL

+++ Assets/DLYH/Scripts/Networking/LocalAIOpponent.cs
@@ -119,7 +119,7 @@ namespace DLYH.Networking
 
             // Calculate miss limit based on player's grid settings
             _missLimit = GameplayStateTracker.CalculateMissLimit(
-                localPlayerSetup.DifficultyLevel,
+                (int)localPlayerSetup.DifficultyLevel,
                 localPlayerSetup.GridSize,
                 localPlayerSetup.WordCount
             );
